ATR-855: added cache attached event info to event infos hierarchy
