+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/SemanticModel/AcumaticaEvents/Base/GraphFieldEventInfoBase.cs
@@ -1,0 +1,104 @@
+﻿using System;
+
+using Acuminator.Utilities.Common;
+using Acuminator.Utilities.Roslyn.Semantic.Dac;
+
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+
+namespace Acuminator.Utilities.Roslyn.Semantic.AcumaticaEvents
+{
+	/// <summary>
+	/// A base type for information about the graph event handlers related to DAC fields.
+	/// </summary>
+	public abstract class GraphFieldEventInfoBase<TFieldEventHandlerInfo> : GraphEventHandlerInfoBase<TFieldEventHandlerInfo>
+	where TFieldEventHandlerInfo : GraphFieldEventInfoBase<TFieldEventHandlerInfo>
+	{
+		/// <summary>
+		/// The DAC field name.
+		/// </summary>
+		public string DacFieldName { get; }
+
+		public GraphFieldEventInfoBase(MethodDeclarationSyntax? node, IMethodSymbol symbol, int declarationOrder,
+									   EventHandlerLooseInfo handlerLooseInfo) :
+								  base(node, symbol, declarationOrder, handlerLooseInfo)
+		{
+			DacFieldName = GetDacFieldName();
+		}
+
+		public GraphFieldEventInfoBase(MethodDeclarationSyntax? node, IMethodSymbol symbol, int declarationOrder,
+									   EventHandlerLooseInfo handlerLooseInfo, TFieldEventHandlerInfo baseInfo) : 
+								  base(node, symbol, declarationOrder, handlerLooseInfo, baseInfo)
+		{
+			DacFieldName = GetDacFieldName();
+		}
+
+		internal override string GetEventGroupingKey() => $"{DacName}_{DacFieldName}_{EventType.ToString()}";
+
+		private string GetDacFieldName() =>
+			SignatureType switch
+			{
+				EventHandlerSignatureType.Classic => GetDacFieldNameForDefaultFieldEvent(),
+				EventHandlerSignatureType.Generic => GetDacFieldNameForGenericFieldEvent(),
+				_ => string.Empty,
+			};
+
+		private string GetDacFieldNameForDefaultFieldEvent()
+		{
+			string eventName = Symbol.Name;
+
+			if (eventName.Length < 5 || eventName[0] == '_' || eventName[^1] == '_')
+			{
+				return string.Empty;
+			}
+
+			int underscoreCount = 0;
+			int firstUnderscoreIndex = -1, lastUnderscoreIndex = -1;
+
+			for (int i = 0; i < eventName.Length; i++)
+			{
+				if (eventName[i] != '_')
+					continue;
+
+				if (underscoreCount == 0)
+				{
+					firstUnderscoreIndex = i;
+				}
+				else
+				{
+					lastUnderscoreIndex = i;
+				}
+
+				underscoreCount++;
+			}
+
+			if (underscoreCount != 2 || firstUnderscoreIndex < 0 || lastUnderscoreIndex < 0)
+				return string.Empty;
+
+			int length = lastUnderscoreIndex - firstUnderscoreIndex - 1;
+
+			if (length == 0)
+				return string.Empty;
+
+			return eventName.Substring(firstUnderscoreIndex + 1, length);
+		}
+
+		private string GetDacFieldNameForGenericFieldEvent()
+		{
+			if (Symbol.Parameters.IsDefaultOrEmpty || Symbol.Parameters[0]?.Type is not INamedTypeSymbol firstParameter)
+			{
+				return string.Empty;
+			}
+
+			ITypeSymbol? dacField = firstParameter.TypeArguments.Length > 1
+				? firstParameter.TypeArguments[1]
+				: firstParameter.TypeArguments.Length == 1
+					? firstParameter.TypeArguments[0]
+					: null;
+
+			return dacField != null && dacField.IsDacBqlField()
+					? dacField.Name.ToPascalCase()
+					: string.Empty;
+		}	
+	}
+}
