+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/SemanticModel/AcumaticaEvents/GraphFieldEventInfo.cs
@@ -1,9 +1,4 @@
-﻿#nullable enable
-
-using System;
-
-using Acuminator.Utilities.Common;
-using Acuminator.Utilities.Roslyn.Semantic.Dac;
+﻿using System;
 
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -13,99 +8,29 @@ namespace Acuminator.Utilities.Roslyn.Semantic.AcumaticaEvents
 	/// <summary>
 	/// Information about the graph event handlers related to the DAC fields.
 	/// </summary>
-	public class GraphFieldEventInfo : GraphEventHandlerInfoBase<GraphFieldEventInfo>
+	public class GraphFieldEventInfo : GraphFieldEventInfoBase<GraphFieldEventInfo>
 	{
-		/// <summary>
-		/// The DAC field name.
-		/// </summary>
-		public string DacFieldName { get; }
-
 		public GraphFieldEventInfo(MethodDeclarationSyntax? node, IMethodSymbol symbol, int declarationOrder,
 								   EventHandlerLooseInfo handlerLooseInfo) :
 							  base(node, symbol, declarationOrder, handlerLooseInfo)
 		{
-			DacFieldName = GetDacFieldName();
 		}
 
 		public GraphFieldEventInfo(MethodDeclarationSyntax? node, IMethodSymbol symbol, int declarationOrder,
-								   EventHandlerLooseInfo handlerLooseInfo, GraphFieldEventInfo baseInfo)
-							: base(node, symbol, declarationOrder, handlerLooseInfo, baseInfo)
+								   EventHandlerLooseInfo handlerLooseInfo, GraphFieldEventInfo baseInfo) :
+							  base(node, symbol, declarationOrder, handlerLooseInfo, baseInfo)
 		{
-			DacFieldName = GetDacFieldName();
 		}
-
-		internal override string GetEventGroupingKey() => $"{DacName}_{DacFieldName}_{EventType.ToString()}";
 
 		protected override void ValidateEventType(EventType eventType)
 		{
 			if (!eventType.IsDacFieldEvent())
 				throw new ArgumentOutOfRangeException(nameof(eventType), $"The {eventType} is not a field event type.");
+			else if (eventType == EventType.CacheAttached)
+			{
+				throw new ArgumentOutOfRangeException(nameof(eventType), $"The {eventType} is a cache attached event." +
+													  $"The {nameof(GraphCacheAttachedEventInfo)} type should be used.");
+			}
 		}
-
-		private string GetDacFieldName() =>
-			SignatureType switch
-			{
-				EventHandlerSignatureType.Classic => GetDacFieldNameForDefaultFieldEvent(),
-				EventHandlerSignatureType.Generic => GetDacFieldNameForGenericFieldEvent(),
-				_ => string.Empty,
-			};
-
-		private string GetDacFieldNameForDefaultFieldEvent()
-		{
-			string eventName = Symbol.Name;
-
-			if (eventName.Length < 5 || eventName[0] == '_' || eventName[^1] == '_')
-			{
-				return string.Empty;
-			}
-
-			int underscoreCount = 0;
-			int firstUnderscoreIndex = -1, lastUnderscoreIndex = -1;
-
-			for (int i = 0; i < eventName.Length; i++)
-			{
-				if (eventName[i] != '_')
-					continue;
-
-				if (underscoreCount == 0)
-				{
-					firstUnderscoreIndex = i;
-				}
-				else
-				{
-					lastUnderscoreIndex = i;
-				}
-
-				underscoreCount++;
-			}
-
-			if (underscoreCount != 2 || firstUnderscoreIndex < 0 || lastUnderscoreIndex < 0)
-				return string.Empty;
-
-			int length = lastUnderscoreIndex - firstUnderscoreIndex - 1;
-
-			if (length == 0)
-				return string.Empty;
-
-			return eventName.Substring(firstUnderscoreIndex + 1, length);
-		}
-
-		private string GetDacFieldNameForGenericFieldEvent()
-		{
-			if (Symbol.Parameters.IsDefaultOrEmpty || Symbol.Parameters[0]?.Type is not INamedTypeSymbol firstParameter)
-			{
-				return string.Empty;
-			}
-
-			ITypeSymbol? dacField = firstParameter.TypeArguments.Length > 1
-				? firstParameter.TypeArguments[1]
-				: firstParameter.TypeArguments.Length == 1
-					? firstParameter.TypeArguments[0]
-					: null;
-
-			return dacField != null && dacField.IsDacBqlField()
-					? dacField.Name.ToPascalCase()
-					: string.Empty;
-		}	
 	}
 }
